// src/api/openapi/models/orderableType.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const OrderableType = {
    Job: 'job',
    Stock: 'stock',
    Supplier: 'supplier'
} as const;
export type OrderableType = typeof OrderableType[keyof typeof OrderableType];